requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.22
pyyaml>=6.0
rank-bm25>=0.2.2
numpy>=1.24
requests>=2.31.0
whoosh>=2.7.4
markdown>=3.4.1
python-dotenv>=1.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
python-dotenv>=1.0.0
requests>=2.31.0
whoosh>=2.7.4
markdown>=3.4.1
pydantic>=2.0.0